+++ db/sql/120_staging_schema.sql
@@ -172,7 +172,7 @@ CREATE TABLE staging.apc_raw (
   "start"               interval,
   observed_operator_id  smallint,
   stop                  integer,
-  vehicle_load          smallint,
+  vehicle_load          integer,
   vehicle_load_ratio    real,
   doors_data            jsonb,
   count_quality         text,
